Deprecate (remove) Repository non-async methods
