+++ src/SourceCode.Chasm.IO.Azure.Tests/ChasmRepositoryTests.cs
@@ -1,5 +1,6 @@
 using Moq;
 using SourceCode.Chasm.IO.Json;
+using System.Threading;
 using Xunit;
 
 namespace SourceCode.Chasm.IO.Azure.Tests
@@ -23,13 +24,13 @@ namespace SourceCode.Chasm.IO.Azure.Tests
             repo.Setup(r => r.Serializer).Returns(new JsonChasmSerializer());
             repo.Setup(r => r.CompressionLevel).Returns(System.IO.Compression.CompressionLevel.NoCompression);
 
-            repo.Setup(r => r.ReadCommitRef(null, null)).Returns(CommitId.Empty);
+            repo.Setup(r => r.ReadCommitRefAsync(null, null, CancellationToken.None).Result).Returns(CommitId.Empty);
 
-            repo.Setup(r => r.ReadCommitRef(string.Empty, null)).Returns(CommitId.Empty);
-            repo.Setup(r => r.ReadCommitRef(null, string.Empty)).Returns(CommitId.Empty);
-            repo.Setup(r => r.ReadCommitRef(string.Empty, string.Empty)).Returns(CommitId.Empty);
+            repo.Setup(r => r.ReadCommitRefAsync(string.Empty, null, CancellationToken.None).Result).Returns(CommitId.Empty);
+            repo.Setup(r => r.ReadCommitRefAsync(null, string.Empty, CancellationToken.None).Result).Returns(CommitId.Empty);
+            repo.Setup(r => r.ReadCommitRefAsync(string.Empty, string.Empty, CancellationToken.None).Result).Returns(CommitId.Empty);
 
-            repo.Setup(r => r.ReadCommitRef("branch", "name")).Returns(new CommitId(Sha1.Hash("branch-name")));
+            repo.Setup(r => r.ReadCommitRefAsync("branch", "name", CancellationToken.None).Result).Returns(new CommitId(Sha1.Hash("branch-name")));
         }
     }
 }
