+++ src/SourceCode.Chasm.IO.AzureBlobStorage/AzureBlobChasmRepo.Commit.cs
@@ -1,22 +1,10 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 
-namespace SourceCode.Chasm.IO
+namespace SourceCode.Chasm.IO.AzureBlobStorage
 {
-    partial class ChasmRepository // .Commit
+    partial class AzureBlobChasmRepo // .Commit
     {
-        #region Read
-
-        public Commit ReadCommit(CommitId commitId)
-        {
-            var buffer = ReadObject(commitId.Sha1);
-            if (buffer.IsEmpty)
-                return Commit.Empty;
-
-            var model = Serializer.DeserializeCommit(buffer.Span);
-            return model;
-        }
-
         public async ValueTask<Commit> ReadCommitAsync(CommitId commitId, CancellationToken cancellationToken)
         {
             var buffer = await ReadObjectAsync(commitId.Sha1, cancellationToken).ConfigureAwait(false);
@@ -25,23 +13,6 @@ namespace SourceCode.Chasm.IO
 
             var model = Serializer.DeserializeCommit(buffer.Span);
             return model;
-        }
-
-        #endregion
-
-        #region Write
-
-        public CommitId WriteCommit(Commit commit, bool forceOverwrite)
-        {
-            using (var session = Serializer.Serialize(commit))
-            {
-                var sha1 = Sha1.Hash(session.Result);
-
-                WriteObject(sha1, session.Result, forceOverwrite);
-
-                var commitId = new CommitId(sha1);
-                return commitId;
-            }
         }
 
         public async ValueTask<CommitId> WriteCommitAsync(Commit commit, bool forceOverwrite, CancellationToken cancellationToken)
@@ -56,7 +27,5 @@ namespace SourceCode.Chasm.IO
                 return commitId;
             }
         }
-
-        #endregion
     }
 }
