+++ src/SourceCode.Chasm.IO.AzureBlobStorage/AzureBlobChasmRepo.cs
@@ -6,7 +6,7 @@ using System.Threading;
 
 namespace SourceCode.Chasm.IO.AzureBlobStorage
 {
-    public sealed partial class AzureBlobChasmRepo : ChasmRepository
+    public sealed partial class AzureBlobChasmRepo : IChasmRepository
     {
         #region Fields
 
@@ -15,12 +15,27 @@ namespace SourceCode.Chasm.IO.AzureBlobStorage
 
         #endregion
 
+        #region Properties
+
+        public IChasmSerializer Serializer { get; }
+
+        public CompressionLevel CompressionLevel { get; }
+
+        public int MaxDop { get; }
+
+        #endregion
+
         #region Constructors
 
-        public AzureBlobChasmRepo(ChasmSerializer serializer, CloudStorageAccount storageAccount, CompressionLevel compressionLevel)
-            : base(serializer, compressionLevel)
+        public AzureBlobChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer, CompressionLevel compressionLevel, int maxDop)
         {
             if (storageAccount == null) throw new ArgumentNullException(nameof(storageAccount));
+            if (!Enum.IsDefined(typeof(CompressionLevel), compressionLevel)) throw new ArgumentOutOfRangeException(nameof(compressionLevel));
+            if (maxDop < -1 || maxDop == 0) throw new ArgumentOutOfRangeException(nameof(maxDop));
+
+            Serializer = serializer ?? throw new ArgumentNullException(nameof(serializer));
+            CompressionLevel = compressionLevel;
+            MaxDop = maxDop;
 
             var client = storageAccount.CreateCloudBlobClient();
             client.DefaultRequestOptions.ParallelOperationThreadCount = 4; // Default is 1
@@ -28,7 +43,7 @@ namespace SourceCode.Chasm.IO.AzureBlobStorage
             // Refs
             _refsContainer = new Lazy<CloudBlobContainer>(() =>
             {
-                var container = $"refs";
+                var container = "refs";
                 var tr = client.GetContainerReference(container);
 
                 tr.CreateIfNotExistsAsync().Wait();
@@ -48,27 +63,28 @@ namespace SourceCode.Chasm.IO.AzureBlobStorage
             }, LazyThreadSafetyMode.PublicationOnly);
         }
 
+        public AzureBlobChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer, CompressionLevel compressionLevel)
+          : this(storageAccount, serializer, compressionLevel, -1)
+        { }
+
+        public AzureBlobChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer)
+            : this(storageAccount, serializer, CompressionLevel.Optimal)
+        { }
+
         #endregion
 
         #region Factory
 
-        public static AzureBlobChasmRepo Create(ChasmSerializer serializer, string connectionString, CompressionLevel compressionLevel)
+        public static AzureBlobChasmRepo Create(string connectionString, IChasmSerializer serializer, CompressionLevel compressionLevel)
         {
             if (string.IsNullOrWhiteSpace(connectionString)) throw new ArgumentException(nameof(connectionString));
             if (serializer == null) throw new ArgumentNullException(nameof(serializer));
 
             var storageAccount = CloudStorageAccount.Parse(connectionString);
-            var repo = new AzureBlobChasmRepo(serializer, storageAccount, compressionLevel);
+            var repo = new AzureBlobChasmRepo(storageAccount, serializer, compressionLevel);
 
             return repo;
         }
-
-        #endregion
-
-        #region Helpers
-
-        private static ChasmConcurrencyException BuildConcurrencyException(string branch, string name, Exception innerException)
-            => new ChasmConcurrencyException($"Concurrent write detected on {nameof(CommitRef)} {branch}/{name}", innerException);
 
         #endregion
     }
