+++ src/SourceCode.Chasm.IO.AzureTableStorage/AzureTableChasmRepo.cs
@@ -6,7 +6,7 @@ using System.Threading;
 
 namespace SourceCode.Chasm.IO.AzureTableStorage
 {
-    public sealed partial class AzureTableChasmRepo : ChasmRepository
+    public sealed partial class AzureTableChasmRepo : IChasmRepository
     {
         #region Fields
 
@@ -15,12 +15,27 @@ namespace SourceCode.Chasm.IO.AzureTableStorage
 
         #endregion
 
+        #region Properties
+
+        public IChasmSerializer Serializer { get; }
+
+        public CompressionLevel CompressionLevel { get; }
+
+        public int MaxDop { get; }
+
+        #endregion
+
         #region Constructors
 
-        public AzureTableChasmRepo(ChasmSerializer serializer, CloudStorageAccount storageAccount, CompressionLevel compressionLevel)
-            : base(serializer, compressionLevel)
+        public AzureTableChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer, CompressionLevel compressionLevel, int maxDop)
         {
             if (storageAccount == null) throw new ArgumentNullException(nameof(storageAccount));
+            if (!Enum.IsDefined(typeof(CompressionLevel), compressionLevel)) throw new ArgumentOutOfRangeException(nameof(compressionLevel));
+            if (maxDop < -1 || maxDop == 0) throw new ArgumentOutOfRangeException(nameof(maxDop));
+
+            Serializer = serializer ?? throw new ArgumentNullException(nameof(serializer));
+            CompressionLevel = compressionLevel;
+            MaxDop = maxDop;
 
             var client = storageAccount.CreateCloudTableClient();
 
@@ -47,27 +62,28 @@ namespace SourceCode.Chasm.IO.AzureTableStorage
             }, LazyThreadSafetyMode.PublicationOnly);
         }
 
+        public AzureTableChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer, CompressionLevel compressionLevel)
+          : this(storageAccount, serializer, compressionLevel, -1)
+        { }
+
+        public AzureTableChasmRepo(CloudStorageAccount storageAccount, IChasmSerializer serializer)
+            : this(storageAccount, serializer, CompressionLevel.Optimal)
+        { }
+
         #endregion
 
         #region Factory
 
-        public static AzureTableChasmRepo Create(ChasmSerializer serializer, string connectionString, CompressionLevel compressionLevel)
+        public static AzureTableChasmRepo Create(string connectionString, IChasmSerializer serializer, CompressionLevel compressionLevel)
         {
             if (string.IsNullOrWhiteSpace(connectionString)) throw new ArgumentException(nameof(connectionString));
             if (serializer == null) throw new ArgumentNullException(nameof(serializer));
 
             var storageAccount = CloudStorageAccount.Parse(connectionString);
-            var repo = new AzureTableChasmRepo(serializer, storageAccount, compressionLevel);
+            var repo = new AzureTableChasmRepo(storageAccount, serializer, compressionLevel);
 
             return repo;
         }
-
-        #endregion
-
-        #region Helpers
-
-        private static ChasmConcurrencyException BuildConcurrencyException(string branch, string name, Exception innerException)
-            => new ChasmConcurrencyException($"Concurrent write detected on {nameof(CommitRef)} {branch}/{name}", innerException);
 
         #endregion
     }
