+++ src/SourceCode.Chasm/IO/IChasmRepository.Commit.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace SourceCode.Chasm.IO
@@ -7,15 +7,11 @@ namespace SourceCode.Chasm.IO
     {
         #region Read
 
-        Commit ReadCommit(CommitId commitId);
-
         ValueTask<Commit> ReadCommitAsync(CommitId commitId, CancellationToken cancellationToken);
 
         #endregion
 
         #region Write
-
-        CommitId WriteCommit(Commit commit, bool forceOverwrite);
 
         ValueTask<CommitId> WriteCommitAsync(Commit commit, bool forceOverwrite, CancellationToken cancellationToken);
