+++ src/SourceCode.Chasm/IO/IChasmRepository.CommitRef.cs
@@ -1,4 +1,4 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace SourceCode.Chasm.IO
@@ -7,25 +7,11 @@ namespace SourceCode.Chasm.IO
     {
         #region Read
 
-        CommitId ReadCommitRef(string branch, string name);
-
         ValueTask<CommitId> ReadCommitRefAsync(string branch, string name, CancellationToken cancellationToken);
 
         #endregion
 
         #region Write
-
-        /// <summary>
-        /// Write a <see cref="CommitRef"/> to the repository using the provided values.
-        /// Note that the underlying store should use pessimistic concurrency control to prevent data loss.
-        /// </summary>
-        /// <param name="branch">The repository name.</param>
-        /// <param name="name">The name of the <see cref="CommitRef"/></param>
-        /// <param name="previousCommitId">The previous <see cref="CommitId"/> that the caller used for reading.</param>
-        /// <param name="newCommitId">The new <see cref="CommitId"/> that represents the content being written.</param>
-        /// <exception cref="ChasmConcurrencyException">Thrown when a concurrency exception is detected.</exception>
-        /// <returns></returns>
-        void WriteCommitRef(CommitId? previousCommitId, string branch, string name, CommitId newCommitId);
 
         /// <summary>
         /// Write a <see cref="CommitRef"/> to the repository using the provided values.
