+++ src/SourceCode.Chasm/IO/IChasmRepository.Object.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,11 +9,7 @@ namespace SourceCode.Chasm.IO
     {
         #region Read
 
-        ReadOnlyMemory<byte> ReadObject(Sha1 objectId);
-
         ValueTask<ReadOnlyMemory<byte>> ReadObjectAsync(Sha1 objectId, CancellationToken cancellationToken);
-
-        IReadOnlyDictionary<Sha1, ReadOnlyMemory<byte>> ReadObjects(IEnumerable<Sha1> objectIds, CancellationToken cancellationToken);
 
         ValueTask<IReadOnlyDictionary<Sha1, ReadOnlyMemory<byte>>> ReadObjectsAsync(IEnumerable<Sha1> objectIds, CancellationToken cancellationToken);
 
@@ -21,11 +17,7 @@ namespace SourceCode.Chasm.IO
 
         #region Write
 
-        void WriteObject(Sha1 objectId, ArraySegment<byte> segment, bool forceOverwrite);
-
         Task WriteObjectAsync(Sha1 objectId, ArraySegment<byte> segment, bool forceOverwrite, CancellationToken cancellationToken);
-
-        void WriteObjects(IEnumerable<KeyValuePair<Sha1, ArraySegment<byte>>> items, bool forceOverwrite, CancellationToken cancellationToken);
 
         Task WriteObjectsAsync(IEnumerable<KeyValuePair<Sha1, ArraySegment<byte>>> items, bool forceOverwrite, CancellationToken cancellationToken);
