+++ src/SourceCode.Chasm/IO/IChasmRepository.Tree.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -8,19 +8,13 @@ namespace SourceCode.Chasm.IO
     {
         #region Read
 
-        TreeNodeList ReadTree(TreeId treeId);
-
         ValueTask<TreeNodeList> ReadTreeAsync(TreeId treeId, CancellationToken cancellationToken);
-
-        IReadOnlyDictionary<TreeId, TreeNodeList> ReadTrees(IEnumerable<TreeId> treeIds, CancellationToken cancellationToken);
 
         ValueTask<IReadOnlyDictionary<TreeId, TreeNodeList>> ReadTreesAsync(IEnumerable<TreeId> treeIds, CancellationToken cancellationToken);
 
         #endregion
 
         #region Write
-
-        TreeId WriteTree(TreeNodeList tree, bool forceOverwrite);
 
         ValueTask<TreeId> WriteTreeAsync(TreeNodeList tree, bool forceOverwrite, CancellationToken cancellationToken);
