+++ src/SourceCode.Chasm/SourceCode.Chasm.csproj
@@ -12,8 +12,8 @@
     <PackageReference Include="SonarAnalyzer.CSharp" Version="6.4.1.3596">
       <PrivateAssets>All</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="SourceCode.Clay.Buffers" Version="1.0.0-preview1-00175" />
-    <PackageReference Include="SourceCode.Clay.Collections" Version="1.0.0-preview1-00175" />
+    <PackageReference Include="SourceCode.Clay.Buffers" Version="1.0.0-preview1-00182" />
+    <PackageReference Include="SourceCode.Clay.Collections" Version="1.0.0-preview1-00182" />
   </ItemGroup>
 
   <PropertyGroup>
